+++ backend/api/v1/rollout_service.go
@@ -5,24 +5,18 @@ import (
 	"encoding/json"
 	"fmt"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/pkg/errors"
-	"go.uber.org/zap"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/bytebase/bytebase/backend/common"
-	"github.com/bytebase/bytebase/backend/common/log"
 	"github.com/bytebase/bytebase/backend/component/activity"
 	"github.com/bytebase/bytebase/backend/component/dbfactory"
 	"github.com/bytebase/bytebase/backend/component/state"
 	enterpriseAPI "github.com/bytebase/bytebase/backend/enterprise/api"
 	api "github.com/bytebase/bytebase/backend/legacyapi"
-	"github.com/bytebase/bytebase/backend/plugin/db"
-	parser "github.com/bytebase/bytebase/backend/plugin/parser/sql"
 	"github.com/bytebase/bytebase/backend/runner/plancheck"
 	"github.com/bytebase/bytebase/backend/runner/taskcheck"
 	"github.com/bytebase/bytebase/backend/store"
@@ -617,718 +611,6 @@ func (s *RolloutService) BatchCancelTaskRuns(ctx context.Context, request *v1pb.
 	return &v1pb.BatchCancelTaskRunsResponse{}, nil
 }
 
-func convertToPlanCheckRuns(ctx context.Context, s *store.Store, parent string, runs []*store.PlanCheckRunMessage) ([]*v1pb.PlanCheckRun, error) {
-	var planCheckRuns []*v1pb.PlanCheckRun
-	for _, run := range runs {
-		converted, err := convertToPlanCheckRun(ctx, s, parent, run)
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to convert plan check run")
-		}
-		planCheckRuns = append(planCheckRuns, converted)
-	}
-	return planCheckRuns, nil
-}
-
-func convertToPlanCheckRun(ctx context.Context, s *store.Store, parent string, run *store.PlanCheckRunMessage) (*v1pb.PlanCheckRun, error) {
-	converted := &v1pb.PlanCheckRun{
-		Name:       fmt.Sprintf("%s/%s%d", parent, common.PlanCheckRunPrefix, run.UID),
-		Uid:        fmt.Sprintf("%d", run.UID),
-		CreateTime: timestamppb.New(time.Unix(run.CreatedTs, 0)),
-		Type:       convertToPlanCheckRunType(run.Type),
-		Status:     convertToPlanCheckRunStatus(run.Status),
-		Target:     "",
-		Sheet:      "",
-		Results:    convertToPlanCheckRunResults(run.Result.Results),
-		Error:      run.Result.Error,
-	}
-	if run.Config.DatabaseId != 0 {
-		databaseUID := int(run.Config.DatabaseId)
-		database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{UID: &databaseUID})
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to get database")
-		}
-		converted.Target = fmt.Sprintf("%s%s/%s%s", common.InstanceNamePrefix, database.InstanceID, common.DatabaseIDPrefix, database.DatabaseName)
-	}
-	if run.Config.SheetId != 0 {
-		sheetUID := int(run.Config.SheetId)
-		sheet, err := s.GetSheet(ctx, &store.FindSheetMessage{UID: &sheetUID}, api.SystemBotID)
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to get sheet")
-		}
-		sheetProject, err := s.GetProjectV2(ctx, &store.FindProjectMessage{UID: &sheet.ProjectUID})
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to get sheet project")
-		}
-		converted.Sheet = fmt.Sprintf("%s%s/%s%d", common.ProjectNamePrefix, sheetProject.ResourceID, common.SheetIDPrefix, sheet.UID)
-	}
-	return converted, nil
-}
-
-func convertToPlanCheckRunType(t store.PlanCheckRunType) v1pb.PlanCheckRun_Type {
-	switch t {
-	case store.PlanCheckDatabaseStatementFakeAdvise:
-		return v1pb.PlanCheckRun_DATABASE_STATEMENT_FAKE_ADVISE
-	case store.PlanCheckDatabaseStatementCompatibility:
-		return v1pb.PlanCheckRun_DATABASE_STATEMENT_COMPATIBILITY
-	case store.PlanCheckDatabaseStatementAdvise:
-		return v1pb.PlanCheckRun_DATABASE_STATEMENT_FAKE_ADVISE
-	case store.PlanCheckDatabaseStatementType:
-		return v1pb.PlanCheckRun_DATABASE_STATEMENT_TYPE
-	case store.PlanCheckDatabaseStatementSummaryReport:
-		return v1pb.PlanCheckRun_DATABASE_STATEMENT_SUMMARY_REPORT
-	case store.PlanCheckDatabaseConnect:
-		return v1pb.PlanCheckRun_DATABASE_CONNECT
-	case store.PlanCheckDatabaseGhostSync:
-		return v1pb.PlanCheckRun_DATABASE_GHOST_SYNC
-	case store.PlanCheckDatabasePITRMySQL:
-		return v1pb.PlanCheckRun_DATABASE_PITR_MYSQL
-	}
-	return v1pb.PlanCheckRun_TYPE_UNSPECIFIED
-}
-
-func convertToPlanCheckRunStatus(status store.PlanCheckRunStatus) v1pb.PlanCheckRun_Status {
-	switch status {
-	case store.PlanCheckRunStatusCanceled:
-		return v1pb.PlanCheckRun_CANCELED
-	case store.PlanCheckRunStatusDone:
-		return v1pb.PlanCheckRun_DONE
-	case store.PlanCheckRunStatusFailed:
-		return v1pb.PlanCheckRun_FAILED
-	case store.PlanCheckRunStatusRunning:
-		return v1pb.PlanCheckRun_RUNNING
-	}
-	return v1pb.PlanCheckRun_STATUS_UNSPECIFIED
-}
-
-func convertToPlanCheckRunResults(results []*storepb.PlanCheckRunResult_Result) []*v1pb.PlanCheckRun_Result {
-	var resultsV1 []*v1pb.PlanCheckRun_Result
-	for _, result := range results {
-		resultsV1 = append(resultsV1, convertToPlanCheckRunResult(result))
-	}
-	return resultsV1
-}
-
-func convertToPlanCheckRunResult(result *storepb.PlanCheckRunResult_Result) *v1pb.PlanCheckRun_Result {
-	resultV1 := &v1pb.PlanCheckRun_Result{
-		Status:  convertToPlanCheckRunResultStatus(result.Status),
-		Title:   result.Title,
-		Content: result.Content,
-		Code:    result.Code,
-		Report:  nil,
-	}
-	switch report := result.Report.(type) {
-	case *storepb.PlanCheckRunResult_Result_SqlSummaryReport_:
-		resultV1.Report = &v1pb.PlanCheckRun_Result_SqlSummaryReport_{
-			SqlSummaryReport: &v1pb.PlanCheckRun_Result_SqlSummaryReport{
-				StatementType: report.SqlSummaryReport.StatementType,
-				AffectedRows:  report.SqlSummaryReport.AffectedRows,
-			},
-		}
-	case *storepb.PlanCheckRunResult_Result_SqlReviewReport_:
-		resultV1.Report = &v1pb.PlanCheckRun_Result_SqlReviewReport_{
-			SqlReviewReport: &v1pb.PlanCheckRun_Result_SqlReviewReport{
-				Line:   report.SqlReviewReport.Line,
-				Detail: report.SqlReviewReport.Detail,
-				Code:   report.SqlReviewReport.Code,
-			},
-		}
-	}
-	return resultV1
-}
-
-func convertToPlanCheckRunResultStatus(status storepb.PlanCheckRunResult_Result_Status) v1pb.PlanCheckRun_Result_Status {
-	switch status {
-	case storepb.PlanCheckRunResult_Result_STATUS_UNSPECIFIED:
-		return v1pb.PlanCheckRun_Result_STATUS_UNSPECIFIED
-	case storepb.PlanCheckRunResult_Result_SUCCESS:
-		return v1pb.PlanCheckRun_Result_SUCCESS
-	case storepb.PlanCheckRunResult_Result_WARNING:
-		return v1pb.PlanCheckRun_Result_WARNING
-	case storepb.PlanCheckRunResult_Result_ERROR:
-		return v1pb.PlanCheckRun_Result_ERROR
-	}
-	return v1pb.PlanCheckRun_Result_STATUS_UNSPECIFIED
-}
-
-func convertToTaskRuns(taskRuns []*store.TaskRunMessage) []*v1pb.TaskRun {
-	var taskRunsV1 []*v1pb.TaskRun
-	for _, taskRun := range taskRuns {
-		taskRunsV1 = append(taskRunsV1, convertToTaskRun(taskRun))
-	}
-	return taskRunsV1
-}
-
-func convertToTaskRunStatus(status api.TaskRunStatus) v1pb.TaskRun_Status {
-	switch status {
-	case api.TaskRunUnknown:
-		return v1pb.TaskRun_STATUS_UNSPECIFIED
-	case api.TaskRunPending:
-		return v1pb.TaskRun_PENDING
-	case api.TaskRunRunning:
-		return v1pb.TaskRun_RUNNING
-	case api.TaskRunDone:
-		return v1pb.TaskRun_DONE
-	case api.TaskRunFailed:
-		return v1pb.TaskRun_FAILED
-	case api.TaskRunCanceled:
-		return v1pb.TaskRun_CANCELED
-	default:
-		return v1pb.TaskRun_STATUS_UNSPECIFIED
-	}
-}
-
-func convertToTaskRun(taskRun *store.TaskRunMessage) *v1pb.TaskRun {
-	return &v1pb.TaskRun{
-		Name:          fmt.Sprintf("%s%s/%s%d/%s%d/%s%d/%s%d", common.ProjectNamePrefix, taskRun.ProjectID, common.RolloutPrefix, taskRun.PipelineUID, common.StagePrefix, taskRun.StageUID, common.TaskPrefix, taskRun.TaskUID, common.TaskRunPrefix, taskRun.ID),
-		Uid:           fmt.Sprintf("%d", taskRun.ID),
-		Creator:       fmt.Sprintf("users/%s", taskRun.Creator.Email),
-		Updater:       fmt.Sprintf("users/%s", taskRun.Updater.Email),
-		CreateTime:    timestamppb.New(time.Unix(taskRun.CreatedTs, 0)),
-		UpdateTime:    timestamppb.New(time.Unix(taskRun.UpdatedTs, 0)),
-		Title:         taskRun.Name,
-		Status:        convertToTaskRunStatus(taskRun.Status),
-		Detail:        taskRun.ResultProto.Detail,
-		ChangeHistory: taskRun.ResultProto.ChangeHistory,
-		SchemaVersion: taskRun.ResultProto.Version,
-	}
-}
-
-func convertToRollout(ctx context.Context, s *store.Store, project *store.ProjectMessage, rollout *store.PipelineMessage) (*v1pb.Rollout, error) {
-	rolloutV1 := &v1pb.Rollout{
-		Name:   fmt.Sprintf("%s%s/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, rollout.ID),
-		Uid:    fmt.Sprintf("%d", rollout.ID),
-		Plan:   "",
-		Title:  rollout.Name,
-		Stages: nil,
-	}
-
-	taskIDToName := map[int]string{}
-	for _, stage := range rollout.Stages {
-		environment, err := s.GetEnvironmentV2(ctx, &store.FindEnvironmentMessage{
-			UID: &stage.EnvironmentID,
-		})
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to get environment %d", stage.EnvironmentID)
-		}
-		if environment == nil {
-			return nil, errors.Errorf("environment %d not found", stage.EnvironmentID)
-		}
-		rolloutStage := &v1pb.Stage{
-			Name:        fmt.Sprintf("%s%s/%s%d/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, rollout.ID, common.StagePrefix, stage.ID),
-			Uid:         fmt.Sprintf("%d", stage.ID),
-			Environment: fmt.Sprintf("%s%s", common.EnvironmentNamePrefix, environment.ResourceID),
-			Title:       stage.Name,
-		}
-		for _, task := range stage.TaskList {
-			rolloutTask, err := convertToTask(ctx, s, project, task)
-			if err != nil {
-				return nil, status.Errorf(codes.Internal, "failed to convert task, error: %v", err)
-			}
-			taskIDToName[task.ID] = rolloutTask.Name
-			rolloutStage.Tasks = append(rolloutStage.Tasks, rolloutTask)
-		}
-
-		rolloutV1.Stages = append(rolloutV1.Stages, rolloutStage)
-	}
-
-	for i, rolloutStage := range rolloutV1.Stages {
-		for j, rolloutTask := range rolloutStage.Tasks {
-			task := rollout.Stages[i].TaskList[j]
-			for _, blockingTask := range task.BlockedBy {
-				rolloutTask.BlockedByTasks = append(rolloutTask.BlockedByTasks, taskIDToName[blockingTask])
-			}
-		}
-	}
-
-	return rolloutV1, nil
-}
-
-func convertToTask(ctx context.Context, s *store.Store, project *store.ProjectMessage, task *store.TaskMessage) (*v1pb.Task, error) {
-	switch task.Type {
-	case api.TaskDatabaseCreate:
-		return convertToTaskFromDatabaseCreate(ctx, s, project, task)
-	case api.TaskDatabaseSchemaBaseline:
-		return convertToTaskFromSchemaBaseline(ctx, s, project, task)
-	case api.TaskDatabaseSchemaUpdate, api.TaskDatabaseSchemaUpdateSDL, api.TaskDatabaseSchemaUpdateGhostSync:
-		return convertToTaskFromSchemaUpdate(ctx, s, project, task)
-	case api.TaskDatabaseSchemaUpdateGhostCutover:
-		return convertToTaskFromSchemaUpdateGhostCutover(ctx, s, project, task)
-	case api.TaskDatabaseDataUpdate:
-		return convertToTaskFromDataUpdate(ctx, s, project, task)
-	case api.TaskDatabaseBackup:
-		return convertToTaskFromDatabaseBackUp(ctx, s, project, task)
-	case api.TaskDatabaseRestorePITRRestore:
-		return convertToTaskFromDatabaseRestoreRestore(ctx, s, project, task)
-	case api.TaskDatabaseRestorePITRCutover:
-		return convertToTaskFromDatabaseRestoreCutOver(ctx, s, project, task)
-	case api.TaskGeneral:
-		fallthrough
-	default:
-		return nil, errors.Errorf("task type %v is not supported", task.Type)
-	}
-}
-
-func convertToDatabaseLabels(labelsJSON string) (map[string]string, error) {
-	if labelsJSON == "" {
-		return nil, nil
-	}
-	var labels []*api.DatabaseLabel
-	if err := json.Unmarshal([]byte(labelsJSON), &labels); err != nil {
-		return nil, err
-	}
-	labelsMap := make(map[string]string)
-	for _, label := range labels {
-		labelsMap[label.Key] = label.Value
-	}
-	return labelsMap, nil
-}
-
-func convertToTaskFromDatabaseCreate(ctx context.Context, s *store.Store, project *store.ProjectMessage, task *store.TaskMessage) (*v1pb.Task, error) {
-	payload := &api.TaskDatabaseCreatePayload{}
-	if err := json.Unmarshal([]byte(task.Payload), payload); err != nil {
-		return nil, errors.Wrapf(err, "failed to unmarshal task payload")
-	}
-	instance, err := s.GetInstanceV2(ctx, &store.FindInstanceMessage{
-		UID: &task.InstanceID,
-	})
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get instance %d", task.InstanceID)
-	}
-	labels, err := convertToDatabaseLabels(payload.Labels)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to convert database labels %v", payload.Labels)
-	}
-	v1pbTask := &v1pb.Task{
-		Name:           fmt.Sprintf("%s%s/%s%d/%s%d/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, task.PipelineID, common.StagePrefix, task.StageID, common.TaskPrefix, task.ID),
-		Uid:            fmt.Sprintf("%d", task.ID),
-		Title:          task.Name,
-		SpecId:         payload.SpecID,
-		Type:           convertToTaskType(task.Type),
-		Status:         convertToTaskStatus(task.LatestTaskRunStatus, payload.Skipped),
-		SkippedReason:  payload.SkippedReason,
-		BlockedByTasks: nil,
-		Target:         fmt.Sprintf("%s%s", common.InstanceNamePrefix, instance.ResourceID),
-		Payload: &v1pb.Task_DatabaseCreate_{
-			DatabaseCreate: &v1pb.Task_DatabaseCreate{
-				Project:      "",
-				Database:     payload.DatabaseName,
-				Table:        payload.TableName,
-				Sheet:        getResourceNameForSheet(project, payload.SheetID),
-				CharacterSet: payload.CharacterSet,
-				Collation:    payload.Collation,
-				Labels:       labels,
-			},
-		},
-	}
-
-	return v1pbTask, nil
-}
-
-func convertToTaskFromSchemaBaseline(ctx context.Context, s *store.Store, project *store.ProjectMessage, task *store.TaskMessage) (*v1pb.Task, error) {
-	if task.DatabaseID == nil {
-		return nil, errors.Errorf("database id is nil")
-	}
-	payload := &api.TaskDatabaseSchemaBaselinePayload{}
-	if err := json.Unmarshal([]byte(task.Payload), payload); err != nil {
-		return nil, errors.Wrapf(err, "failed to unmarshal task payload")
-	}
-	database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{UID: task.DatabaseID})
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get database")
-	}
-	if database == nil {
-		return nil, errors.Errorf("database not found")
-	}
-	v1pbTask := &v1pb.Task{
-		Name:           fmt.Sprintf("%s%s/%s%d/%s%d/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, task.PipelineID, common.StagePrefix, task.StageID, common.TaskPrefix, task.ID),
-		Uid:            fmt.Sprintf("%d", task.ID),
-		Title:          task.Name,
-		SpecId:         payload.SpecID,
-		Type:           convertToTaskType(task.Type),
-		Status:         convertToTaskStatus(task.LatestTaskRunStatus, payload.Skipped),
-		SkippedReason:  payload.SkippedReason,
-		BlockedByTasks: nil,
-		Target:         fmt.Sprintf("%s%s/%s%s", common.InstanceNamePrefix, database.InstanceID, common.DatabaseIDPrefix, database.DatabaseName),
-		Payload: &v1pb.Task_DatabaseSchemaBaseline_{
-			DatabaseSchemaBaseline: &v1pb.Task_DatabaseSchemaBaseline{
-				SchemaVersion: payload.SchemaVersion,
-			},
-		},
-	}
-	return v1pbTask, nil
-}
-
-func convertToTaskFromSchemaUpdate(ctx context.Context, s *store.Store, project *store.ProjectMessage, task *store.TaskMessage) (*v1pb.Task, error) {
-	if task.DatabaseID == nil {
-		return nil, errors.Errorf("database id is nil")
-	}
-	payload := &api.TaskDatabaseSchemaUpdatePayload{}
-	if err := json.Unmarshal([]byte(task.Payload), payload); err != nil {
-		return nil, errors.Wrapf(err, "failed to unmarshal task payload")
-	}
-	database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{UID: task.DatabaseID})
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get database")
-	}
-	if database == nil {
-		return nil, errors.Errorf("database not found")
-	}
-
-	// HACK: task.Statement is not empty means that the statement comes from a database group target.
-	// we don't want to create new sheets every time so we pass the statement as sheet.
-	sheet := getResourceNameForSheet(project, payload.SheetID)
-	if task.Statement != "" {
-		sheet = task.Statement
-	}
-
-	v1pbTask := &v1pb.Task{
-		Name:           fmt.Sprintf("%s%s/%s%d/%s%d/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, task.PipelineID, common.StagePrefix, task.StageID, common.TaskPrefix, task.ID),
-		Uid:            fmt.Sprintf("%d", task.ID),
-		Title:          task.Name,
-		SpecId:         payload.SpecID,
-		Type:           convertToTaskType(task.Type),
-		Status:         convertToTaskStatus(task.LatestTaskRunStatus, payload.Skipped),
-		SkippedReason:  payload.SkippedReason,
-		BlockedByTasks: nil,
-		Target:         fmt.Sprintf("%s%s/%s%s", common.InstanceNamePrefix, database.InstanceID, common.DatabaseIDPrefix, database.DatabaseName),
-		Payload: &v1pb.Task_DatabaseSchemaUpdate_{
-			DatabaseSchemaUpdate: &v1pb.Task_DatabaseSchemaUpdate{
-				Sheet:         sheet,
-				SchemaVersion: payload.SchemaVersion,
-			},
-		},
-	}
-	return v1pbTask, nil
-}
-
-func convertToTaskFromSchemaUpdateGhostCutover(ctx context.Context, s *store.Store, project *store.ProjectMessage, task *store.TaskMessage) (*v1pb.Task, error) {
-	if task.DatabaseID == nil {
-		return nil, errors.Errorf("database id is nil")
-	}
-	payload := &api.TaskDatabaseSchemaUpdateGhostCutoverPayload{}
-	if err := json.Unmarshal([]byte(task.Payload), payload); err != nil {
-		return nil, errors.Wrapf(err, "failed to unmarshal task payload")
-	}
-	database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{UID: task.DatabaseID})
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get database")
-	}
-	if database == nil {
-		return nil, errors.Errorf("database not found")
-	}
-	v1pbTask := &v1pb.Task{
-		Name:           fmt.Sprintf("%s%s/%s%d/%s%d/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, task.PipelineID, common.StagePrefix, task.StageID, common.TaskPrefix, task.ID),
-		Uid:            fmt.Sprintf("%d", task.ID),
-		Title:          task.Name,
-		SpecId:         payload.SpecID,
-		Status:         convertToTaskStatus(task.LatestTaskRunStatus, payload.Skipped),
-		SkippedReason:  payload.SkippedReason,
-		Type:           convertToTaskType(task.Type),
-		BlockedByTasks: nil,
-		Target:         fmt.Sprintf("%s%s/%s%s", common.InstanceNamePrefix, database.InstanceID, common.DatabaseIDPrefix, database.DatabaseName),
-		Payload:        nil,
-	}
-	return v1pbTask, nil
-}
-
-func convertToTaskFromDataUpdate(ctx context.Context, s *store.Store, project *store.ProjectMessage, task *store.TaskMessage) (*v1pb.Task, error) {
-	if task.DatabaseID == nil {
-		return nil, errors.Errorf("database id is nil")
-	}
-	payload := &api.TaskDatabaseDataUpdatePayload{}
-	if err := json.Unmarshal([]byte(task.Payload), payload); err != nil {
-		return nil, errors.Wrapf(err, "failed to unmarshal task payload")
-	}
-	var rollbackSheetName string
-	if payload.RollbackSheetID != 0 {
-		rollbackSheetName = getResourceNameForSheet(project, payload.RollbackSheetID)
-	}
-	database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{UID: task.DatabaseID})
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get database")
-	}
-	if database == nil {
-		return nil, errors.Errorf("database not found")
-	}
-
-	// HACK: task.Statement is not empty means that the statement comes from a database group target.
-	// we don't want to create new sheets every time so we pass the statement as sheet.
-	sheet := getResourceNameForSheet(project, payload.SheetID)
-	if task.Statement != "" {
-		sheet = task.Statement
-	}
-
-	v1pbTask := &v1pb.Task{
-		Name:           fmt.Sprintf("%s%s/%s%d/%s%d/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, task.PipelineID, common.StagePrefix, task.StageID, common.TaskPrefix, task.ID),
-		Uid:            fmt.Sprintf("%d", task.ID),
-		Title:          task.Name,
-		SpecId:         payload.SpecID,
-		Type:           convertToTaskType(task.Type),
-		Status:         convertToTaskStatus(task.LatestTaskRunStatus, payload.Skipped),
-		SkippedReason:  payload.SkippedReason,
-		BlockedByTasks: nil,
-		Target:         fmt.Sprintf("%s%s/%s%s", common.InstanceNamePrefix, database.InstanceID, common.DatabaseIDPrefix, database.DatabaseName),
-		Payload:        nil,
-	}
-	v1pbTaskPayload := &v1pb.Task_DatabaseDataUpdate_{
-		DatabaseDataUpdate: &v1pb.Task_DatabaseDataUpdate{
-			Sheet:             sheet,
-			SchemaVersion:     payload.SchemaVersion,
-			RollbackEnabled:   payload.RollbackEnabled,
-			RollbackSqlStatus: convertToRollbackSQLStatus(payload.RollbackSQLStatus),
-			RollbackError:     payload.RollbackError,
-			RollbackSheet:     rollbackSheetName,
-			RollbackFromIssue: "",
-			RollbackFromTask:  "",
-		},
-	}
-	if payload.RollbackFromIssueID != 0 && payload.RollbackFromTaskID != 0 {
-		rollbackFromIssue, err := s.GetIssueV2(ctx, &store.FindIssueMessage{
-			UID: &payload.RollbackFromIssueID,
-		})
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to get rollback issue %q", payload.RollbackFromIssueID)
-		}
-		rollbackFromTask, err := s.GetTaskV2ByID(ctx, payload.RollbackFromTaskID)
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to get rollback task %q", payload.RollbackFromTaskID)
-		}
-		v1pbTaskPayload.DatabaseDataUpdate.RollbackFromIssue = fmt.Sprintf("%s%s/%s%d", common.ProjectNamePrefix, project.ResourceID, common.IssuePrefix, rollbackFromIssue.UID)
-		v1pbTaskPayload.DatabaseDataUpdate.RollbackFromTask = fmt.Sprintf("%s%s/%s%d/%s%d/%s%d", common.ProjectNamePrefix, rollbackFromIssue.Project.ResourceID, common.RolloutPrefix, rollbackFromTask.PipelineID, common.StagePrefix, rollbackFromTask.StageID, common.TaskPrefix, rollbackFromTask.ID)
-	}
-
-	v1pbTask.Payload = v1pbTaskPayload
-	return v1pbTask, nil
-}
-
-func convertToTaskFromDatabaseBackUp(ctx context.Context, s *store.Store, project *store.ProjectMessage, task *store.TaskMessage) (*v1pb.Task, error) {
-	if task.DatabaseID == nil {
-		return nil, errors.Errorf("database id is nil")
-	}
-	payload := &api.TaskDatabaseBackupPayload{}
-	if err := json.Unmarshal([]byte(task.Payload), payload); err != nil {
-		return nil, errors.Wrapf(err, "failed to unmarshal task payload")
-	}
-	backup, err := s.GetBackupByUID(ctx, payload.BackupID)
-	if err != nil {
-		return nil, errors.Errorf("failed to get backup by uid: %v", err)
-	}
-	if backup == nil {
-		return nil, errors.Errorf("backup not found")
-	}
-	databaseBackup, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{
-		UID: &backup.DatabaseUID,
-	})
-	if err != nil {
-		return nil, errors.Errorf("failed to get database: %v", err)
-	}
-	if databaseBackup == nil {
-		return nil, errors.Errorf("database not found")
-	}
-	database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{UID: task.DatabaseID})
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get database")
-	}
-	if database == nil {
-		return nil, errors.Errorf("database not found")
-	}
-	v1pbTask := &v1pb.Task{
-		Name:           fmt.Sprintf("%s%s/%s%d/%s%d/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, task.PipelineID, common.StagePrefix, task.StageID, common.TaskPrefix, task.ID),
-		Uid:            fmt.Sprintf("%d", task.ID),
-		Title:          task.Name,
-		SpecId:         payload.SpecID,
-		Type:           convertToTaskType(task.Type),
-		Status:         convertToTaskStatus(task.LatestTaskRunStatus, payload.Skipped),
-		SkippedReason:  payload.SkippedReason,
-		BlockedByTasks: nil,
-		Target:         fmt.Sprintf("%s%s/%s%s", common.InstanceNamePrefix, database.InstanceID, common.DatabaseIDPrefix, database.DatabaseName),
-		Payload: &v1pb.Task_DatabaseBackup_{
-			DatabaseBackup: &v1pb.Task_DatabaseBackup{
-				Backup: fmt.Sprintf("%s%s/%s%s/%s%d", common.InstanceNamePrefix, databaseBackup.InstanceID, common.DatabaseIDPrefix, databaseBackup.DatabaseName, common.BackupPrefix, backup.UID),
-			},
-		},
-	}
-	return v1pbTask, nil
-}
-
-func convertToTaskFromDatabaseRestoreRestore(ctx context.Context, s *store.Store, project *store.ProjectMessage, task *store.TaskMessage) (*v1pb.Task, error) {
-	if task.DatabaseID == nil {
-		return nil, errors.Errorf("database id is nil")
-	}
-	payload := &api.TaskDatabasePITRRestorePayload{}
-	if err := json.Unmarshal([]byte(task.Payload), payload); err != nil {
-		return nil, errors.Wrapf(err, "failed to unmarshal task payload")
-	}
-	database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{UID: task.DatabaseID})
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get database")
-	}
-	if database == nil {
-		return nil, errors.Errorf("database not found")
-	}
-	v1pbTaskPayload := v1pb.Task_DatabaseRestoreRestore_{
-		DatabaseRestoreRestore: &v1pb.Task_DatabaseRestoreRestore{},
-	}
-	v1pbTask := &v1pb.Task{
-		Name:           fmt.Sprintf("%s%s/%s%d/%s%d/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, task.PipelineID, common.StagePrefix, task.StageID, common.TaskPrefix, task.ID),
-		Uid:            fmt.Sprintf("%d", task.ID),
-		Title:          task.Name,
-		SpecId:         payload.SpecID,
-		Type:           convertToTaskType(task.Type),
-		Status:         convertToTaskStatus(task.LatestTaskRunStatus, payload.Skipped),
-		SkippedReason:  payload.SkippedReason,
-		BlockedByTasks: nil,
-		Target:         fmt.Sprintf("%s%s/%s%s", common.InstanceNamePrefix, database.InstanceID, common.DatabaseIDPrefix, database.DatabaseName),
-		Payload:        nil,
-	}
-	if (payload.BackupID == nil) == (payload.PointInTimeTs == nil) {
-		return nil, errors.Errorf("payload.BackupID and payload.PointInTimeTs cannot be both nil or both not nil")
-	}
-	if (payload.TargetInstanceID == nil) != (payload.DatabaseName == nil) {
-		return nil, errors.Errorf("payload.TargetInstanceID and payload.DatabaseName must be both nil or both not nil")
-	}
-
-	if payload.TargetInstanceID != nil {
-		targetInstance, err := s.GetInstanceV2(ctx, &store.FindInstanceMessage{
-			UID: payload.TargetInstanceID,
-		})
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to get target instance")
-		}
-		if targetInstance == nil {
-			return nil, errors.Errorf("target instance not found")
-		}
-		v1pbTaskPayload.DatabaseRestoreRestore.Target = fmt.Sprintf("%s%s/%s%s", common.InstanceNamePrefix, targetInstance.ResourceID, common.DatabaseIDPrefix, *payload.DatabaseName)
-	}
-
-	if payload.BackupID != nil {
-		backup, err := s.GetBackupByUID(ctx, *payload.BackupID)
-		if err != nil {
-			return nil, errors.Errorf("failed to get backup by uid: %v", err)
-		}
-		if backup == nil {
-			return nil, errors.Errorf("backup not found")
-		}
-		databaseBackup, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{
-			UID: &backup.DatabaseUID,
-		})
-		if err != nil {
-			return nil, errors.Errorf("failed to get database: %v", err)
-		}
-		if databaseBackup == nil {
-			return nil, errors.Errorf("database not found")
-		}
-		v1pbTaskPayload.DatabaseRestoreRestore.Source = &v1pb.Task_DatabaseRestoreRestore_Backup{
-			Backup: fmt.Sprintf("%s%s/%s%s/%s%d", common.InstanceNamePrefix, databaseBackup.InstanceID, common.DatabaseIDPrefix, databaseBackup.DatabaseName, common.BackupPrefix, backup.UID),
-		}
-	}
-	if payload.PointInTimeTs != nil {
-		v1pbTaskPayload.DatabaseRestoreRestore.Source = &v1pb.Task_DatabaseRestoreRestore_PointInTime{
-			PointInTime: timestamppb.New(time.Unix(*payload.PointInTimeTs, 0)),
-		}
-	}
-	v1pbTask.Payload = &v1pbTaskPayload
-
-	return v1pbTask, nil
-}
-
-func convertToTaskFromDatabaseRestoreCutOver(ctx context.Context, s *store.Store, project *store.ProjectMessage, task *store.TaskMessage) (*v1pb.Task, error) {
-	if task.DatabaseID == nil {
-		return nil, errors.Errorf("database id is nil")
-	}
-	payload := &api.TaskDatabasePITRCutoverPayload{}
-	if err := json.Unmarshal([]byte(task.Payload), payload); err != nil {
-		return nil, errors.Wrapf(err, "failed to unmarshal task payload")
-	}
-	database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{UID: task.DatabaseID})
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get database")
-	}
-	if database == nil {
-		return nil, errors.Errorf("database not found")
-	}
-	v1pbTask := &v1pb.Task{
-		Name:           fmt.Sprintf("%s%s/%s%d/%s%d/%s%d", common.ProjectNamePrefix, project.ResourceID, common.RolloutPrefix, task.PipelineID, common.StagePrefix, task.StageID, common.TaskPrefix, task.ID),
-		Uid:            fmt.Sprintf("%d", task.ID),
-		Title:          task.Name,
-		SpecId:         payload.SpecID,
-		Type:           convertToTaskType(task.Type),
-		Status:         convertToTaskStatus(task.LatestTaskRunStatus, payload.Skipped),
-		SkippedReason:  payload.SkippedReason,
-		BlockedByTasks: nil,
-		Target:         fmt.Sprintf("%s%s/%s%s", common.InstanceNamePrefix, database.InstanceID, common.DatabaseIDPrefix, database.DatabaseName),
-		Payload:        nil,
-	}
-
-	return v1pbTask, nil
-}
-
-func convertToTaskStatus(latestTaskRunStatus *api.TaskRunStatus, skipped bool) v1pb.Task_Status {
-	if skipped {
-		return v1pb.Task_SKIPPED
-	}
-	if latestTaskRunStatus == nil {
-		return v1pb.Task_NOT_STARTED
-	}
-	switch *latestTaskRunStatus {
-	case api.TaskRunPending:
-		return v1pb.Task_PENDING
-	case api.TaskRunRunning:
-		return v1pb.Task_RUNNING
-	case api.TaskRunDone:
-		return v1pb.Task_DONE
-	case api.TaskRunFailed:
-		return v1pb.Task_FAILED
-	case api.TaskRunCanceled:
-		return v1pb.Task_CANCELED
-	default:
-		return v1pb.Task_STATUS_UNSPECIFIED
-	}
-}
-
-func convertToTaskType(taskType api.TaskType) v1pb.Task_Type {
-	switch taskType {
-	case api.TaskGeneral:
-		return v1pb.Task_GENERAL
-	case api.TaskDatabaseCreate:
-		return v1pb.Task_DATABASE_CREATE
-	case api.TaskDatabaseSchemaBaseline:
-		return v1pb.Task_DATABASE_SCHEMA_BASELINE
-	case api.TaskDatabaseSchemaUpdate:
-		return v1pb.Task_DATABASE_SCHEMA_UPDATE
-	case api.TaskDatabaseSchemaUpdateSDL:
-		return v1pb.Task_DATABASE_SCHEMA_UPDATE_SDL
-	case api.TaskDatabaseSchemaUpdateGhostSync:
-		return v1pb.Task_DATABASE_SCHEMA_UPDATE_GHOST_SYNC
-	case api.TaskDatabaseSchemaUpdateGhostCutover:
-		return v1pb.Task_DATABASE_SCHEMA_UPDATE_GHOST_CUTOVER
-	case api.TaskDatabaseDataUpdate:
-		return v1pb.Task_DATABASE_DATA_UPDATE
-	case api.TaskDatabaseBackup:
-		return v1pb.Task_DATABASE_BACKUP
-	case api.TaskDatabaseRestorePITRRestore:
-		return v1pb.Task_DATABASE_RESTORE_RESTORE
-	case api.TaskDatabaseRestorePITRCutover:
-		return v1pb.Task_DATABASE_RESTORE_CUTOVER
-	default:
-		return v1pb.Task_TYPE_UNSPECIFIED
-	}
-}
-
-func convertToRollbackSQLStatus(status api.RollbackSQLStatus) v1pb.Task_DatabaseDataUpdate_RollbackSqlStatus {
-	switch status {
-	case api.RollbackSQLStatusPending:
-		return v1pb.Task_DatabaseDataUpdate_PENDING
-	case api.RollbackSQLStatusDone:
-		return v1pb.Task_DatabaseDataUpdate_DONE
-	case api.RollbackSQLStatusFailed:
-		return v1pb.Task_DatabaseDataUpdate_FAILED
-
-	default:
-		return v1pb.Task_DatabaseDataUpdate_ROLLBACK_SQL_STATUS_UNSPECIFIED
-	}
-}
-
 // UpdatePlan updates a plan.
 // Currently, only Spec.Config.Sheet can be updated.
 func (s *RolloutService) UpdatePlan(ctx context.Context, request *v1pb.UpdatePlanRequest) (*v1pb.Plan, error) {
@@ -1550,1310 +832,6 @@ func (s *RolloutService) getPipelineCreate(ctx context.Context, steps []*storepb
 	return pipelineCreate, nil
 }
 
-func (s *RolloutService) getTaskCreatesFromSpec(ctx context.Context, spec *storepb.PlanConfig_Spec, project *store.ProjectMessage, registerEnvironmentID func(string) error) ([]*store.TaskMessage, []store.TaskIndexDAG, error) {
-	if s.licenseService.IsFeatureEnabled(api.FeatureTaskScheduleTime) != nil {
-		if spec.EarliestAllowedTime != nil && !spec.EarliestAllowedTime.AsTime().IsZero() {
-			return nil, nil, errors.Errorf(api.FeatureTaskScheduleTime.AccessErrorMessage())
-		}
-	}
-
-	switch config := spec.Config.(type) {
-	case *storepb.PlanConfig_Spec_CreateDatabaseConfig:
-		return getTaskCreatesFromCreateDatabaseConfig(ctx, s.store, s.licenseService, s.dbFactory, spec, config.CreateDatabaseConfig, project, registerEnvironmentID)
-	case *storepb.PlanConfig_Spec_ChangeDatabaseConfig:
-		return getTaskCreatesFromChangeDatabaseConfig(ctx, s.store, spec, config.ChangeDatabaseConfig, project, registerEnvironmentID)
-	case *storepb.PlanConfig_Spec_RestoreDatabaseConfig:
-		return getTaskCreatesFromRestoreDatabaseConfig(ctx, s.store, s.licenseService, s.dbFactory, spec, config.RestoreDatabaseConfig, project, registerEnvironmentID)
-	}
-
-	return nil, nil, errors.Errorf("invalid spec config type %T", spec.Config)
-}
-
-func getTaskCreatesFromCreateDatabaseConfig(ctx context.Context, s *store.Store, licenseService enterpriseAPI.LicenseService, dbFactory *dbfactory.DBFactory, spec *storepb.PlanConfig_Spec, c *storepb.PlanConfig_CreateDatabaseConfig, project *store.ProjectMessage, registerEnvironmentID func(string) error) ([]*store.TaskMessage, []store.TaskIndexDAG, error) {
-	if c.Database == "" {
-		return nil, nil, errors.Errorf("database name is required")
-	}
-	instanceID, err := common.GetInstanceID(c.Target)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get instance id from %q", c.Target)
-	}
-	instance, err := s.GetInstanceV2(ctx, &store.FindInstanceMessage{ResourceID: &instanceID})
-	if err != nil {
-		return nil, nil, err
-	}
-	if instance == nil {
-		return nil, nil, errors.Errorf("instance ID not found %v", instanceID)
-	}
-	if instance.Engine == db.Oracle {
-		return nil, nil, errors.Errorf("creating Oracle database is not supported")
-	}
-	environment, err := s.GetEnvironmentV2(ctx, &store.FindEnvironmentMessage{ResourceID: &instance.EnvironmentID})
-	if err != nil {
-		return nil, nil, err
-	}
-	if environment == nil {
-		return nil, nil, errors.Errorf("environment ID not found %v", instance.EnvironmentID)
-	}
-
-	if err := registerEnvironmentID(environment.ResourceID); err != nil {
-		return nil, nil, err
-	}
-
-	if instance.Engine == db.MongoDB && c.Table == "" {
-		return nil, nil, errors.Errorf("collection name is required for MongoDB")
-	}
-
-	taskCreates, err := func() ([]*store.TaskMessage, error) {
-		if err := checkCharacterSetCollationOwner(instance.Engine, c.CharacterSet, c.Collation, c.Owner); err != nil {
-			return nil, err
-		}
-		if c.Database == "" {
-			return nil, errors.Errorf("database name is required")
-		}
-		if instance.Engine == db.Snowflake {
-			// Snowflake needs to use upper case of DatabaseName.
-			c.Database = strings.ToUpper(c.Database)
-		}
-		if instance.Engine == db.MongoDB && c.Table == "" {
-			return nil, common.Errorf(common.Invalid, "Failed to create issue, collection name missing for MongoDB")
-		}
-		// Validate the labels. Labels are set upon task completion.
-		labelsJSON, err := convertDatabaseLabels(c.Labels)
-		if err != nil {
-			return nil, errors.Wrapf(err, "invalid database label %q", c.Labels)
-		}
-
-		// We will use schema from existing tenant databases for creating a database in a tenant mode project if possible.
-		if project.TenantMode == api.TenantModeTenant {
-			if err := licenseService.IsFeatureEnabled(api.FeatureMultiTenancy); err != nil {
-				return nil, err
-			}
-		}
-
-		// Get admin data source username.
-		adminDataSource := utils.DataSourceFromInstanceWithType(instance, api.Admin)
-		if adminDataSource == nil {
-			return nil, common.Errorf(common.Internal, "admin data source not found for instance %q", instance.Title)
-		}
-		databaseName := c.Database
-		switch instance.Engine {
-		case db.Snowflake:
-			// Snowflake needs to use upper case of DatabaseName.
-			databaseName = strings.ToUpper(databaseName)
-		case db.MySQL, db.MariaDB, db.OceanBase:
-			// For MySQL, we need to use different case of DatabaseName depends on the variable `lower_case_table_names`.
-			// https://dev.mysql.com/doc/refman/8.0/en/identifier-case-sensitivity.html
-			// And also, meet an error in here is not a big deal, we will just use the original DatabaseName.
-			driver, err := dbFactory.GetAdminDatabaseDriver(ctx, instance, nil /* database */)
-			if err != nil {
-				log.Warn("failed to get admin database driver for instance %q, please check the connection for admin data source", zap.Error(err), zap.String("instance", instance.Title))
-				break
-			}
-			defer driver.Close(ctx)
-			var lowerCaseTableNames int
-			var unused any
-			db := driver.GetDB()
-			if err := db.QueryRowContext(ctx, "SHOW VARIABLES LIKE 'lower_case_table_names'").Scan(&unused, &lowerCaseTableNames); err != nil {
-				log.Warn("failed to get lower_case_table_names for instance %q", zap.Error(err), zap.String("instance", instance.Title))
-				break
-			}
-			if lowerCaseTableNames == 1 {
-				databaseName = strings.ToLower(databaseName)
-			}
-		}
-
-		statement, err := getCreateDatabaseStatement(instance.Engine, c, databaseName, adminDataSource.Username)
-		if err != nil {
-			return nil, err
-		}
-		sheet, err := s.CreateSheet(ctx, &store.SheetMessage{
-			CreatorID:  api.SystemBotID,
-			ProjectUID: project.UID,
-			Name:       fmt.Sprintf("Sheet for creating database %v", databaseName),
-			Statement:  statement,
-			Visibility: store.ProjectSheet,
-			Source:     store.SheetFromBytebaseArtifact,
-			Type:       store.SheetForSQL,
-			Payload:    "{}",
-		})
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to create database creation sheet")
-		}
-
-		payload := api.TaskDatabaseCreatePayload{
-			SpecID:       spec.Id,
-			ProjectID:    project.UID,
-			CharacterSet: c.CharacterSet,
-			TableName:    c.Table,
-			Collation:    c.Collation,
-			Labels:       labelsJSON,
-			DatabaseName: databaseName,
-			SheetID:      sheet.UID,
-		}
-		bytes, err := json.Marshal(payload)
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to create database creation task, unable to marshal payload")
-		}
-
-		return []*store.TaskMessage{
-			{
-				InstanceID:        instance.UID,
-				DatabaseID:        nil,
-				Name:              fmt.Sprintf("Create database %v", payload.DatabaseName),
-				Status:            api.TaskPendingApproval,
-				Type:              api.TaskDatabaseCreate,
-				DatabaseName:      payload.DatabaseName,
-				Payload:           string(bytes),
-				EarliestAllowedTs: spec.EarliestAllowedTime.GetSeconds(),
-			},
-		}, nil
-	}()
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return taskCreates, nil, nil
-}
-
-func getTaskCreatesFromChangeDatabaseConfig(ctx context.Context, s *store.Store, spec *storepb.PlanConfig_Spec, c *storepb.PlanConfig_ChangeDatabaseConfig, project *store.ProjectMessage, registerEnvironmentID func(string) error) ([]*store.TaskMessage, []store.TaskIndexDAG, error) {
-	// possible target:
-	// 1. instances/{instance}/databases/{database}
-	// 2. projects/{project}/databaseGroups/{databaseGroup}
-	if _, _, err := common.GetInstanceDatabaseID(c.Target); err == nil {
-		return getTaskCreatesFromChangeDatabaseConfigDatabaseTarget(ctx, s, spec, c, project, registerEnvironmentID)
-	}
-	if _, _, err := common.GetProjectIDDatabaseGroupID(c.Target); err == nil {
-		return getTaskCreatesFromChangeDatabaseConfigDatabaseGroupTarget(ctx, s, spec, c, project, registerEnvironmentID)
-	}
-	return nil, nil, errors.Errorf("unknown target %q", c.Target)
-}
-
-func getTaskCreatesFromChangeDatabaseConfigDatabaseTarget(ctx context.Context, s *store.Store, spec *storepb.PlanConfig_Spec, c *storepb.PlanConfig_ChangeDatabaseConfig, _ *store.ProjectMessage, registerEnvironmentID func(string) error) ([]*store.TaskMessage, []store.TaskIndexDAG, error) {
-	instanceID, databaseName, err := common.GetInstanceDatabaseID(c.Target)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get instance and database from target %q", c.Target)
-	}
-
-	instance, err := s.GetInstanceV2(ctx, &store.FindInstanceMessage{
-		ResourceID: &instanceID,
-	})
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get instance %q", instanceID)
-	}
-	if instance == nil {
-		return nil, nil, errors.Errorf("instance %q not found", instanceID)
-	}
-	database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{
-		InstanceID:          &instanceID,
-		DatabaseName:        &databaseName,
-		IgnoreCaseSensitive: store.IgnoreDatabaseAndTableCaseSensitive(instance),
-	})
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get database %q", databaseName)
-	}
-	if database == nil {
-		return nil, nil, errors.Errorf("database %q not found", databaseName)
-	}
-
-	if err := registerEnvironmentID(database.EffectiveEnvironmentID); err != nil {
-		return nil, nil, err
-	}
-
-	switch c.Type {
-	case storepb.PlanConfig_ChangeDatabaseConfig_BASELINE:
-		payload := api.TaskDatabaseSchemaBaselinePayload{
-			SpecID:        spec.Id,
-			SchemaVersion: getOrDefaultSchemaVersion(c.SchemaVersion),
-		}
-		bytes, err := json.Marshal(payload)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to marshal task database schema baseline payload")
-		}
-		payloadString := string(bytes)
-		taskCreate := &store.TaskMessage{
-			Name:              fmt.Sprintf("Establish baseline for database %q", database.DatabaseName),
-			InstanceID:        instance.UID,
-			DatabaseID:        &database.UID,
-			Status:            api.TaskPendingApproval,
-			Type:              api.TaskDatabaseSchemaBaseline,
-			EarliestAllowedTs: spec.EarliestAllowedTime.GetSeconds(),
-			Payload:           payloadString,
-		}
-		return []*store.TaskMessage{taskCreate}, nil, nil
-
-	case storepb.PlanConfig_ChangeDatabaseConfig_MIGRATE:
-		_, sheetIDStr, err := common.GetProjectResourceIDSheetID(c.Sheet)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to get sheet id from sheet %q", c.Sheet)
-		}
-		sheetID, err := strconv.Atoi(sheetIDStr)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to convert sheet id %q to int", sheetIDStr)
-		}
-		payload := api.TaskDatabaseSchemaUpdatePayload{
-			SpecID:        spec.Id,
-			SheetID:       sheetID,
-			SchemaVersion: getOrDefaultSchemaVersion(c.SchemaVersion),
-			VCSPushEvent:  nil,
-		}
-		bytes, err := json.Marshal(payload)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to marshal task database schema update payload")
-		}
-		payloadString := string(bytes)
-		taskCreate := &store.TaskMessage{
-			Name:              fmt.Sprintf("DDL(schema) for database %q", database.DatabaseName),
-			InstanceID:        instance.UID,
-			DatabaseID:        &database.UID,
-			Status:            api.TaskPendingApproval,
-			Type:              api.TaskDatabaseSchemaUpdate,
-			EarliestAllowedTs: spec.EarliestAllowedTime.GetSeconds(),
-			Payload:           payloadString,
-		}
-		return []*store.TaskMessage{taskCreate}, nil, nil
-
-	case storepb.PlanConfig_ChangeDatabaseConfig_MIGRATE_SDL:
-		_, sheetIDStr, err := common.GetProjectResourceIDSheetID(c.Sheet)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to get sheet id from sheet %q", c.Sheet)
-		}
-		sheetID, err := strconv.Atoi(sheetIDStr)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to convert sheet id %q to int", sheetIDStr)
-		}
-		payload := api.TaskDatabaseSchemaUpdateSDLPayload{
-			SpecID:        spec.Id,
-			SheetID:       sheetID,
-			SchemaVersion: getOrDefaultSchemaVersion(c.SchemaVersion),
-			VCSPushEvent:  nil,
-		}
-		bytes, err := json.Marshal(payload)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to marshal database schema update SDL payload")
-		}
-		payloadString := string(bytes)
-		taskCreate := &store.TaskMessage{
-			Name:              fmt.Sprintf("SDL for database %q", database.DatabaseName),
-			InstanceID:        instance.UID,
-			DatabaseID:        &database.UID,
-			Status:            api.TaskPendingApproval,
-			Type:              api.TaskDatabaseSchemaUpdateSDL,
-			EarliestAllowedTs: spec.EarliestAllowedTime.GetSeconds(),
-			Payload:           payloadString,
-		}
-		return []*store.TaskMessage{taskCreate}, nil, nil
-
-	case storepb.PlanConfig_ChangeDatabaseConfig_MIGRATE_GHOST:
-		_, sheetIDStr, err := common.GetProjectResourceIDSheetID(c.Sheet)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to get sheet id from sheet %q", c.Sheet)
-		}
-		sheetID, err := strconv.Atoi(sheetIDStr)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to convert sheet id %q to int", sheetIDStr)
-		}
-		var taskCreateList []*store.TaskMessage
-		// task "sync"
-		payloadSync := api.TaskDatabaseSchemaUpdateGhostSyncPayload{
-			SpecID:        spec.Id,
-			SheetID:       sheetID,
-			SchemaVersion: c.SchemaVersion,
-			VCSPushEvent:  nil,
-		}
-		bytesSync, err := json.Marshal(payloadSync)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to marshal database schema update gh-ost sync payload")
-		}
-		taskCreateList = append(taskCreateList, &store.TaskMessage{
-			Name:              fmt.Sprintf("Update schema gh-ost sync for database %q", database.DatabaseName),
-			InstanceID:        instance.UID,
-			DatabaseID:        &database.UID,
-			Status:            api.TaskPendingApproval,
-			Type:              api.TaskDatabaseSchemaUpdateGhostSync,
-			EarliestAllowedTs: spec.EarliestAllowedTime.GetSeconds(),
-			Payload:           string(bytesSync),
-		})
-
-		// task "cutover"
-		payloadCutover := api.TaskDatabaseSchemaUpdateGhostCutoverPayload{
-			SpecID: spec.Id,
-		}
-		bytesCutover, err := json.Marshal(payloadCutover)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to marshal database schema update ghost cutover payload")
-		}
-		taskCreateList = append(taskCreateList, &store.TaskMessage{
-			Name:              fmt.Sprintf("Update schema gh-ost cutover for database %q", database.DatabaseName),
-			InstanceID:        instance.UID,
-			DatabaseID:        &database.UID,
-			Status:            api.TaskPendingApproval,
-			Type:              api.TaskDatabaseSchemaUpdateGhostCutover,
-			EarliestAllowedTs: spec.EarliestAllowedTime.GetSeconds(),
-			Payload:           string(bytesCutover),
-		})
-
-		// The below list means that taskCreateList[0] blocks taskCreateList[1].
-		// In other words, task "sync" blocks task "cutover".
-		taskIndexDAGList := []store.TaskIndexDAG{
-			{FromIndex: 0, ToIndex: 1},
-		}
-		return taskCreateList, taskIndexDAGList, nil
-
-	case storepb.PlanConfig_ChangeDatabaseConfig_DATA:
-		_, sheetIDStr, err := common.GetProjectResourceIDSheetID(c.Sheet)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to get sheet id from sheet %q", c.Sheet)
-		}
-		sheetID, err := strconv.Atoi(sheetIDStr)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to convert sheet id %q to int", sheetIDStr)
-		}
-		payload := api.TaskDatabaseDataUpdatePayload{
-			SpecID:            spec.Id,
-			SheetID:           sheetID,
-			SchemaVersion:     getOrDefaultSchemaVersion(c.SchemaVersion),
-			VCSPushEvent:      nil,
-			RollbackEnabled:   c.RollbackEnabled,
-			RollbackSQLStatus: api.RollbackSQLStatusPending,
-		}
-		if c.RollbackDetail != nil {
-			issueID, err := common.GetIssueID(c.RollbackDetail.RollbackFromIssue)
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to get issue id from issue %q", c.RollbackDetail.RollbackFromIssue)
-			}
-			payload.RollbackFromIssueID = issueID
-			taskID, err := common.GetTaskID(c.RollbackDetail.RollbackFromTask)
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to get task id from task %q", c.RollbackDetail.RollbackFromTask)
-			}
-			payload.RollbackFromTaskID = taskID
-		}
-		bytes, err := json.Marshal(payload)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "Failed to marshal database data update payload")
-		}
-		payloadString := string(bytes)
-		taskCreate := &store.TaskMessage{
-			Name:              fmt.Sprintf("DML(data) for database %q", database.DatabaseName),
-			InstanceID:        instance.UID,
-			DatabaseID:        &database.UID,
-			Status:            api.TaskPendingApproval,
-			Type:              api.TaskDatabaseDataUpdate,
-			EarliestAllowedTs: spec.EarliestAllowedTime.GetSeconds(),
-			Payload:           payloadString,
-		}
-		return []*store.TaskMessage{taskCreate}, nil, nil
-	default:
-		return nil, nil, errors.Errorf("unsupported change database config type %q", c.Type)
-	}
-}
-
-func getTaskCreatesFromChangeDatabaseConfigDatabaseGroupTarget(ctx context.Context, s *store.Store, spec *storepb.PlanConfig_Spec, c *storepb.PlanConfig_ChangeDatabaseConfig, project *store.ProjectMessage, registerEnvironmentID func(string) error) ([]*store.TaskMessage, []store.TaskIndexDAG, error) {
-	switch c.Type {
-	case storepb.PlanConfig_ChangeDatabaseConfig_MIGRATE:
-	case storepb.PlanConfig_ChangeDatabaseConfig_DATA:
-	default:
-		return nil, nil, errors.Errorf("unsupported change database config type %q for database group target", c.Type)
-	}
-
-	projectID, databaseGroupID, err := common.GetProjectIDDatabaseGroupID(c.Target)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get project id and database group id from target %q", c.Target)
-	}
-	if projectID != project.ResourceID {
-		return nil, nil, errors.Errorf("project id %q in target %q does not match project id %q in plan config", projectID, c.Target, project.ResourceID)
-	}
-	databaseGroup, err := s.GetDatabaseGroup(ctx, &store.FindDatabaseGroupMessage{ProjectUID: &project.UID, ResourceID: &databaseGroupID})
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get database group %q", databaseGroupID)
-	}
-	if databaseGroup == nil {
-		return nil, nil, errors.Errorf("database group %q not found", databaseGroupID)
-	}
-	schemaGroups, err := s.ListSchemaGroups(ctx, &store.FindSchemaGroupMessage{DatabaseGroupUID: &databaseGroup.UID})
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to list schema groups for database group %q", databaseGroupID)
-	}
-	allDatabases, err := s.ListDatabases(ctx, &store.FindDatabaseMessage{ProjectID: &project.ResourceID})
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to list databases for project %q", project.ResourceID)
-	}
-
-	matchedDatabases, _, err := getMatchedAndUnmatchedDatabasesInDatabaseGroup(ctx, databaseGroup, allDatabases)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get matched and unmatched databases in database group %q", databaseGroupID)
-	}
-	if len(matchedDatabases) == 0 {
-		return nil, nil, errors.Errorf("no matched databases found in database group %q", databaseGroupID)
-	}
-
-	var environmentID string
-	for _, db := range matchedDatabases {
-		if environmentID == "" {
-			environmentID = db.EffectiveEnvironmentID
-		}
-		if environmentID != db.EffectiveEnvironmentID {
-			return nil, nil, errors.Errorf("matched databases are in different environments")
-		}
-	}
-	if err := registerEnvironmentID(environmentID); err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to register environment id %q", environmentID)
-	}
-
-	_, sheetUIDStr, err := common.GetProjectResourceIDSheetID(c.Sheet)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get sheet id from sheet %q", c.Sheet)
-	}
-	sheetUID, err := strconv.Atoi(sheetUIDStr)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to convert sheet id %q to int", sheetUIDStr)
-	}
-	sheetStatement, err := s.GetSheetStatementByID(ctx, sheetUID)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get sheet statement %q", sheetUID)
-	}
-
-	var taskCreatesMatrix [][]*store.TaskMessage
-	var taskIndexDAGsMatrix [][]store.TaskIndexDAG
-	for _, db := range matchedDatabases {
-		dbSchema, err := s.GetDBSchema(ctx, db.UID)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to get db schema %q", db.UID)
-		}
-		instance, err := s.GetInstanceV2(ctx, &store.FindInstanceMessage{ResourceID: &db.InstanceID})
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to get instance %q", db.InstanceID)
-		}
-		if instance == nil {
-			return nil, nil, errors.Errorf("instance %q not found", db.InstanceID)
-		}
-
-		schemaGroupsMatchedTables := map[string][]string{}
-		for _, schemaGroup := range schemaGroups {
-			matches, _, err := getMatchedAndUnmatchedTablesInSchemaGroup(ctx, dbSchema, schemaGroup)
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to get matched and unmatched tables in schema group %q", schemaGroup.ResourceID)
-			}
-			schemaGroupsMatchedTables[schemaGroup.ResourceID] = matches
-		}
-
-		parserEngineType, err := convertDatabaseToParserEngineType(instance.Engine)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to convert database engine %q to parser engine type", instance.Engine)
-		}
-
-		statements, schemaGroupNames, err := getStatementsAndSchemaGroupsFromSchemaGroups(sheetStatement, parserEngineType, c.Target, schemaGroups, schemaGroupsMatchedTables)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to get statements from schema groups")
-		}
-
-		taskCreates, err := getTaskCreatesFromChangeDatabaseConfigDatabaseGroupStatements(db, instance, spec, c, statements, schemaGroupNames)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to get task creates from change database config database group statements")
-		}
-		var taskIndexDAGs []store.TaskIndexDAG
-		for i := 1; i < len(taskCreates); i++ {
-			taskIndexDAGs = append(taskIndexDAGs, store.TaskIndexDAG{
-				FromIndex: i - 1,
-				ToIndex:   i,
-			})
-		}
-
-		taskCreatesMatrix = append(taskCreatesMatrix, taskCreates)
-		taskIndexDAGsMatrix = append(taskIndexDAGsMatrix, taskIndexDAGs)
-	}
-
-	creates, dags, err := utils.MergeTaskCreateLists(taskCreatesMatrix, taskIndexDAGsMatrix)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to merge task create lists")
-	}
-
-	return creates, dags, nil
-}
-
-func getTaskCreatesFromChangeDatabaseConfigDatabaseGroupStatements(db *store.DatabaseMessage, instance *store.InstanceMessage, spec *storepb.PlanConfig_Spec, c *storepb.PlanConfig_ChangeDatabaseConfig, statements []string, schemaGroupNames []string) ([]*store.TaskMessage, error) {
-	var creates []*store.TaskMessage
-	for idx, statement := range statements {
-		schemaVersionSuffix := fmt.Sprintf("-%03d", idx)
-		schemaGroupName := schemaGroupNames[idx]
-		switch c.Type {
-		case storepb.PlanConfig_ChangeDatabaseConfig_MIGRATE:
-			payload := api.TaskDatabaseSchemaUpdatePayload{
-				SpecID:          spec.Id,
-				SheetID:         0,
-				SchemaVersion:   getOrDefaultSchemaVersionWithSuffix(c.SchemaVersion, schemaVersionSuffix),
-				VCSPushEvent:    nil,
-				SchemaGroupName: schemaGroupName,
-			}
-			bytes, err := json.Marshal(payload)
-			if err != nil {
-				return nil, errors.Wrapf(err, "failed to marshal task database schema update payload")
-			}
-			payloadString := string(bytes)
-			taskCreate := &store.TaskMessage{
-				Name:              fmt.Sprintf("DDL(schema) for database %q", db.DatabaseName),
-				InstanceID:        instance.UID,
-				DatabaseID:        &db.UID,
-				Status:            api.TaskPendingApproval,
-				Type:              api.TaskDatabaseSchemaUpdate,
-				EarliestAllowedTs: spec.EarliestAllowedTime.GetSeconds(),
-				Payload:           payloadString,
-				Statement:         statement,
-			}
-			creates = append(creates, taskCreate)
-
-		case storepb.PlanConfig_ChangeDatabaseConfig_DATA:
-			payload := api.TaskDatabaseDataUpdatePayload{
-				SpecID:            spec.Id,
-				SheetID:           0,
-				SchemaVersion:     getOrDefaultSchemaVersionWithSuffix(c.SchemaVersion, schemaVersionSuffix),
-				VCSPushEvent:      nil,
-				RollbackEnabled:   c.RollbackEnabled,
-				RollbackSQLStatus: api.RollbackSQLStatusPending,
-				SchemaGroupName:   schemaGroupName,
-			}
-
-			bytes, err := json.Marshal(payload)
-			if err != nil {
-				return nil, errors.Wrapf(err, "failed to marshal database data update payload")
-			}
-			payloadString := string(bytes)
-			taskCreate := &store.TaskMessage{
-				Name:              fmt.Sprintf("DML(data) for database %q", db.DatabaseName),
-				InstanceID:        instance.UID,
-				DatabaseID:        &db.UID,
-				Status:            api.TaskPendingApproval,
-				Type:              api.TaskDatabaseDataUpdate,
-				EarliestAllowedTs: spec.EarliestAllowedTime.GetSeconds(),
-				Payload:           payloadString,
-				Statement:         statement,
-			}
-			creates = append(creates, taskCreate)
-		}
-	}
-
-	return creates, nil
-}
-
-// input: statement, parserEngineType
-// output: rendered statement list
-func getStatementsAndSchemaGroupsFromSchemaGroups(statement string, parserEngineType parser.EngineType, schemaGroupParent string, schemaGroups []*store.SchemaGroupMessage, schemaGroupMatchedTables map[string][]string) ([]string, []string, error) {
-	flush := func(emptyStatementBuilder *strings.Builder, statementBuilder *strings.Builder, schemaGroup *store.SchemaGroupMessage, matchedTables []string) ([]string, []string) {
-		if statementBuilder.Len() == 0 {
-			return nil, nil
-		}
-		var resultStatements, schemaGroupNames []string
-		if len(matchedTables) > 0 {
-			for _, tableName := range matchedTables {
-				statement := emptyStatementBuilder.String() +
-					strings.ReplaceAll(statementBuilder.String(), schemaGroup.Placeholder, tableName)
-				resultStatements = append(resultStatements, statement)
-				schemaGroupNames = append(schemaGroupNames, fmt.Sprintf("%s/%s%s", schemaGroupParent, common.SchemaGroupNamePrefix, schemaGroup.ResourceID))
-			}
-		} else {
-			statement := emptyStatementBuilder.String() + statementBuilder.String()
-			resultStatements = append(resultStatements, statement)
-			schemaGroupNames = append(schemaGroupNames, "")
-		}
-		emptyStatementBuilder.Reset()
-		statementBuilder.Reset()
-		return resultStatements, schemaGroupNames
-	}
-
-	singleStatements, err := parser.SplitMultiSQL(parserEngineType, statement)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to split sql")
-	}
-	if len(singleStatements) == 0 {
-		return nil, nil, errors.Errorf("no sql statement found")
-	}
-
-	var resultStatements, resultSchemaGroupNames []string
-	var emptyStatementBuilder, statementBuilder strings.Builder
-
-	var preMatch, curMatch *store.SchemaGroupMessage
-	for _, singleStatement := range singleStatements {
-		if singleStatement.Empty {
-			_, _ = emptyStatementBuilder.WriteString(singleStatement.Text)
-			continue
-		}
-		for _, schemaGroup := range schemaGroups {
-			if strings.Contains(singleStatement.Text, schemaGroup.Placeholder) {
-				curMatch = schemaGroup
-				break
-			}
-		}
-
-		// discard statement that matches the placeholder but has no matched tables
-		if curMatch != nil && len(schemaGroupMatchedTables[curMatch.ResourceID]) == 0 {
-			curMatch = nil
-			continue
-		}
-
-		if preMatch == nil && curMatch != nil {
-			statements, schemaGroupNames := flush(&emptyStatementBuilder, &statementBuilder, nil, nil)
-			resultStatements = append(resultStatements, statements...)
-			resultSchemaGroupNames = append(resultSchemaGroupNames, schemaGroupNames...)
-		}
-		if preMatch != nil && curMatch == nil {
-			statements, schemaGroupNames := flush(&emptyStatementBuilder, &statementBuilder, preMatch, schemaGroupMatchedTables[preMatch.ResourceID])
-			resultStatements = append(resultStatements, statements...)
-			resultSchemaGroupNames = append(resultSchemaGroupNames, schemaGroupNames...)
-		}
-		if preMatch != nil && curMatch != nil && preMatch.ResourceID != curMatch.ResourceID {
-			statements, schemaGroupNames := flush(&emptyStatementBuilder, &statementBuilder, preMatch, schemaGroupMatchedTables[preMatch.ResourceID])
-			resultStatements = append(resultStatements, statements...)
-			resultSchemaGroupNames = append(resultSchemaGroupNames, schemaGroupNames...)
-		}
-
-		_, _ = statementBuilder.WriteString(singleStatement.Text)
-		_, _ = statementBuilder.WriteString("\n")
-
-		preMatch = curMatch
-		curMatch = nil
-	}
-
-	if preMatch != nil {
-		statements, schemaGroupNames := flush(&emptyStatementBuilder, &statementBuilder, preMatch, schemaGroupMatchedTables[preMatch.ResourceID])
-		resultStatements = append(resultStatements, statements...)
-		resultSchemaGroupNames = append(resultSchemaGroupNames, schemaGroupNames...)
-	} else {
-		statements, schemaGroupNames := flush(&emptyStatementBuilder, &statementBuilder, nil, nil)
-		resultStatements = append(resultStatements, statements...)
-		resultSchemaGroupNames = append(resultSchemaGroupNames, schemaGroupNames...)
-	}
-
-	if emptyStatementBuilder.Len() > 0 && len(resultStatements) > 0 {
-		resultStatements[len(resultStatements)-1] += emptyStatementBuilder.String()
-	}
-
-	return resultStatements, resultSchemaGroupNames, nil
-}
-
-func convertDatabaseToParserEngineType(engine db.Type) (parser.EngineType, error) {
-	switch engine {
-	case db.Oracle:
-		return parser.Oracle, nil
-	case db.MSSQL:
-		return parser.MSSQL, nil
-	case db.Postgres:
-		return parser.Postgres, nil
-	case db.Redshift:
-		return parser.Redshift, nil
-	case db.MySQL:
-		return parser.MySQL, nil
-	case db.TiDB:
-		return parser.TiDB, nil
-	case db.MariaDB:
-		return parser.MariaDB, nil
-	case db.OceanBase:
-		return parser.OceanBase, nil
-	}
-	return parser.EngineType("UNKNOWN"), errors.Errorf("unsupported engine type %q", engine)
-}
-
-func getTaskCreatesFromRestoreDatabaseConfig(ctx context.Context, s *store.Store, licenseService enterpriseAPI.LicenseService, dbFactory *dbfactory.DBFactory, spec *storepb.PlanConfig_Spec, c *storepb.PlanConfig_RestoreDatabaseConfig, project *store.ProjectMessage, registerEnvironmentID func(string) error) ([]*store.TaskMessage, []store.TaskIndexDAG, error) {
-	if c.Source == nil {
-		return nil, nil, errors.Errorf("missing source in restore database config")
-	}
-	instanceID, databaseName, err := common.GetInstanceDatabaseID(c.Target)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get instance and database id from target %q", c.Target)
-	}
-	instance, err := s.GetInstanceV2(ctx, &store.FindInstanceMessage{ResourceID: &instanceID})
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get instance %q", instanceID)
-	}
-	if instance == nil {
-		return nil, nil, errors.Errorf("instance %q not found", instanceID)
-	}
-	database, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{
-		InstanceID:          &instanceID,
-		DatabaseName:        &databaseName,
-		IgnoreCaseSensitive: store.IgnoreDatabaseAndTableCaseSensitive(instance),
-	})
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to get database %q", databaseName)
-	}
-	if database == nil {
-		return nil, nil, errors.Errorf("database %q not found", databaseName)
-	}
-	if database.ProjectID != project.ResourceID {
-		return nil, nil, errors.Errorf("database %q is not in project %q", databaseName, project.ResourceID)
-	}
-
-	if err := registerEnvironmentID(database.EffectiveEnvironmentID); err != nil {
-		return nil, nil, err
-	}
-
-	var taskCreates []*store.TaskMessage
-
-	if c.CreateDatabaseConfig != nil {
-		restorePayload := api.TaskDatabasePITRRestorePayload{
-			SpecID:    spec.Id,
-			ProjectID: project.UID,
-		}
-		// restore to a new database
-		targetInstanceID, err := common.GetInstanceID(c.CreateDatabaseConfig.Target)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to get instance id from %q", c.CreateDatabaseConfig.Target)
-		}
-		targetInstance, err := s.GetInstanceV2(ctx, &store.FindInstanceMessage{ResourceID: &targetInstanceID})
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "failed to find the instance with ID %q", targetInstanceID)
-		}
-
-		// task 1: create the database
-		createDatabaseTasks, _, err := getTaskCreatesFromCreateDatabaseConfig(ctx, s, licenseService, dbFactory, spec, c.CreateDatabaseConfig, project, registerEnvironmentID)
-		if err != nil {
-			return nil, nil, errors.Wrap(err, "failed to create the database create task list")
-		}
-		if len(createDatabaseTasks) != 1 {
-			return nil, nil, errors.Errorf("expected 1 task to create the database, got %d", len(createDatabaseTasks))
-		}
-		taskCreates = append(taskCreates, createDatabaseTasks[0])
-
-		// task 2: restore the database
-		switch source := c.Source.(type) {
-		case *storepb.PlanConfig_RestoreDatabaseConfig_Backup:
-			backupInstanceID, backupDatabaseName, backupName, err := common.GetInstanceDatabaseIDBackupName(source.Backup)
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to parse backup name %q", source.Backup)
-			}
-			backupInstance, err := s.GetInstanceV2(ctx, &store.FindInstanceMessage{ResourceID: &backupInstanceID})
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to get instance %q", backupInstanceID)
-			}
-			backupDatabase, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{
-				InstanceID:          &backupInstanceID,
-				DatabaseName:        &backupDatabaseName,
-				IgnoreCaseSensitive: store.IgnoreDatabaseAndTableCaseSensitive(backupInstance),
-			})
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to get database %q", backupDatabaseName)
-			}
-			if backupDatabase == nil {
-				return nil, nil, errors.Errorf("failed to find database %q where backup %q is created", backupDatabaseName, source.Backup)
-			}
-			backup, err := s.GetBackupV2(ctx, &store.FindBackupMessage{
-				DatabaseUID: &backupDatabase.UID,
-				Name:        &backupName,
-			})
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to get backup %q", backupName)
-			}
-			if backup == nil {
-				return nil, nil, errors.Errorf("failed to find backup %q", backupName)
-			}
-			restorePayload.BackupID = &backup.UID
-		case *storepb.PlanConfig_RestoreDatabaseConfig_PointInTime:
-			ts := source.PointInTime.GetSeconds()
-			restorePayload.PointInTimeTs = &ts
-		}
-		restorePayload.TargetInstanceID = &targetInstance.UID
-		restorePayload.DatabaseName = &c.CreateDatabaseConfig.Database
-
-		restorePayloadBytes, err := json.Marshal(restorePayload)
-		if err != nil {
-			return nil, nil, errors.Wrap(err, "failed to create PITR restore task, unable to marshal payload")
-		}
-
-		restoreTaskCreate := &store.TaskMessage{
-			Name:       fmt.Sprintf("Restore to new database %q", *restorePayload.DatabaseName),
-			Status:     api.TaskPendingApproval,
-			Type:       api.TaskDatabaseRestorePITRRestore,
-			InstanceID: instance.UID,
-			DatabaseID: &database.UID,
-			Payload:    string(restorePayloadBytes),
-		}
-		taskCreates = append(taskCreates, restoreTaskCreate)
-	} else {
-		// in-place restore
-
-		// task 1: restore
-		restorePayload := api.TaskDatabasePITRRestorePayload{
-			SpecID:    spec.Id,
-			ProjectID: project.UID,
-		}
-		switch source := c.Source.(type) {
-		case *storepb.PlanConfig_RestoreDatabaseConfig_Backup:
-			backupInstanceID, backupDatabaseName, backupName, err := common.GetInstanceDatabaseIDBackupName(source.Backup)
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to parse backup name %q", source.Backup)
-			}
-			backupInstance, err := s.GetInstanceV2(ctx, &store.FindInstanceMessage{ResourceID: &backupInstanceID})
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to get instance %q", backupInstanceID)
-			}
-			backupDatabase, err := s.GetDatabaseV2(ctx, &store.FindDatabaseMessage{
-				InstanceID:          &backupInstanceID,
-				DatabaseName:        &backupDatabaseName,
-				IgnoreCaseSensitive: store.IgnoreDatabaseAndTableCaseSensitive(backupInstance),
-			})
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to get database %q", backupDatabaseName)
-			}
-			if backupDatabase == nil {
-				return nil, nil, errors.Errorf("failed to find database %q where backup %q is created", backupDatabaseName, source.Backup)
-			}
-			backup, err := s.GetBackupV2(ctx, &store.FindBackupMessage{
-				DatabaseUID: &backupDatabase.UID,
-				Name:        &backupName,
-			})
-			if err != nil {
-				return nil, nil, errors.Wrapf(err, "failed to get backup %q", backupName)
-			}
-			if backup == nil {
-				return nil, nil, errors.Errorf("failed to find backup %q", backupName)
-			}
-			restorePayload.BackupID = &backup.UID
-		case *storepb.PlanConfig_RestoreDatabaseConfig_PointInTime:
-			ts := source.PointInTime.GetSeconds()
-			restorePayload.PointInTimeTs = &ts
-		}
-		restorePayloadBytes, err := json.Marshal(restorePayload)
-		if err != nil {
-			return nil, nil, errors.Wrap(err, "failed to create PITR restore task, unable to marshal payload")
-		}
-
-		restoreTaskCreate := &store.TaskMessage{
-			Name:       fmt.Sprintf("Restore to PITR database %q", database.DatabaseName),
-			Status:     api.TaskPendingApproval,
-			Type:       api.TaskDatabaseRestorePITRRestore,
-			InstanceID: instance.UID,
-			DatabaseID: &database.UID,
-			Payload:    string(restorePayloadBytes),
-		}
-
-		taskCreates = append(taskCreates, restoreTaskCreate)
-
-		// task 2: cutover
-		cutoverPayload := api.TaskDatabasePITRCutoverPayload{
-			SpecID: spec.Id,
-		}
-		cutoverPayloadBytes, err := json.Marshal(cutoverPayload)
-		if err != nil {
-			return nil, nil, errors.Wrap(err, "failed to create PITR cutover task, unable to marshal payload")
-		}
-		taskCreates = append(taskCreates, &store.TaskMessage{
-			Name:       fmt.Sprintf("Swap PITR and the original database %q", database.DatabaseName),
-			InstanceID: instance.UID,
-			DatabaseID: &database.UID,
-			Status:     api.TaskPendingApproval,
-			Type:       api.TaskDatabaseRestorePITRCutover,
-			Payload:    string(cutoverPayloadBytes),
-		})
-	}
-
-	// We make sure that we will always return 2 tasks.
-	taskIndexDAGs := []store.TaskIndexDAG{
-		{
-			FromIndex: 0,
-			ToIndex:   1,
-		},
-	}
-	return taskCreates, taskIndexDAGs, nil
-}
-
-func convertToPlans(plans []*store.PlanMessage) []*v1pb.Plan {
-	v1Plans := make([]*v1pb.Plan, len(plans))
-	for i := range plans {
-		v1Plans[i] = convertToPlan(plans[i])
-	}
-	return v1Plans
-}
-
-func convertToPlan(plan *store.PlanMessage) *v1pb.Plan {
-	return &v1pb.Plan{
-		Name:        fmt.Sprintf("%s%s/%s%d", common.ProjectNamePrefix, plan.ProjectID, common.PlanPrefix, plan.UID),
-		Uid:         fmt.Sprintf("%d", plan.UID),
-		Issue:       "",
-		Title:       plan.Name,
-		Description: plan.Description,
-		Steps:       convertToPlanSteps(plan.Config.Steps),
-	}
-}
-
-func convertToPlanSteps(steps []*storepb.PlanConfig_Step) []*v1pb.Plan_Step {
-	v1Steps := make([]*v1pb.Plan_Step, len(steps))
-	for i := range steps {
-		v1Steps[i] = convertToPlanStep(steps[i])
-	}
-	return v1Steps
-}
-
-func convertToPlanStep(step *storepb.PlanConfig_Step) *v1pb.Plan_Step {
-	return &v1pb.Plan_Step{
-		Specs: convertToPlanSpecs(step.Specs),
-	}
-}
-
-func convertToPlanSpecs(specs []*storepb.PlanConfig_Spec) []*v1pb.Plan_Spec {
-	v1Specs := make([]*v1pb.Plan_Spec, len(specs))
-	for i := range specs {
-		v1Specs[i] = convertToPlanSpec(specs[i])
-	}
-	return v1Specs
-}
-
-func convertToPlanSpec(spec *storepb.PlanConfig_Spec) *v1pb.Plan_Spec {
-	v1Spec := &v1pb.Plan_Spec{
-		EarliestAllowedTime: spec.EarliestAllowedTime,
-		Id:                  spec.Id,
-	}
-
-	switch v := spec.Config.(type) {
-	case *storepb.PlanConfig_Spec_CreateDatabaseConfig:
-		v1Spec.Config = convertToPlanSpecCreateDatabaseConfig(v)
-	case *storepb.PlanConfig_Spec_ChangeDatabaseConfig:
-		v1Spec.Config = convertToPlanSpecChangeDatabaseConfig(v)
-	case *storepb.PlanConfig_Spec_RestoreDatabaseConfig:
-		v1Spec.Config = convertToPlanSpecRestoreDatabaseConfig(v)
-	}
-
-	return v1Spec
-}
-
-func convertToPlanSpecCreateDatabaseConfig(config *storepb.PlanConfig_Spec_CreateDatabaseConfig) *v1pb.Plan_Spec_CreateDatabaseConfig {
-	c := config.CreateDatabaseConfig
-	return &v1pb.Plan_Spec_CreateDatabaseConfig{
-		CreateDatabaseConfig: &v1pb.Plan_CreateDatabaseConfig{
-			Target:       c.Target,
-			Database:     c.Database,
-			Table:        c.Table,
-			CharacterSet: c.CharacterSet,
-			Collation:    c.Collation,
-			Cluster:      c.Cluster,
-			Owner:        c.Owner,
-			Backup:       c.Backup,
-			Labels:       c.Labels,
-		},
-	}
-}
-
-func convertToPlanCreateDatabaseConfig(c *storepb.PlanConfig_CreateDatabaseConfig) *v1pb.Plan_CreateDatabaseConfig {
-	// c.CreateDatabaseConfig is defined as optional in proto
-	// so we need to test if it's nil
-	if c == nil {
-		return nil
-	}
-	return &v1pb.Plan_CreateDatabaseConfig{
-		Target:       c.Target,
-		Database:     c.Database,
-		Table:        c.Table,
-		CharacterSet: c.CharacterSet,
-		Collation:    c.Collation,
-		Cluster:      c.Cluster,
-		Owner:        c.Owner,
-		Backup:       c.Backup,
-		Labels:       c.Labels,
-	}
-}
-
-func convertToPlanSpecChangeDatabaseConfig(config *storepb.PlanConfig_Spec_ChangeDatabaseConfig) *v1pb.Plan_Spec_ChangeDatabaseConfig {
-	c := config.ChangeDatabaseConfig
-	return &v1pb.Plan_Spec_ChangeDatabaseConfig{
-		ChangeDatabaseConfig: &v1pb.Plan_ChangeDatabaseConfig{
-			Target:          c.Target,
-			Sheet:           c.Sheet,
-			Type:            convertToPlanSpecChangeDatabaseConfigType(c.Type),
-			SchemaVersion:   c.SchemaVersion,
-			RollbackEnabled: c.RollbackEnabled,
-			RollbackDetail:  convertToPlanSpecChangeDatabaseConfigRollbackDetail(c.RollbackDetail),
-		},
-	}
-}
-
-func convertToPlanSpecChangeDatabaseConfigRollbackDetail(d *storepb.PlanConfig_ChangeDatabaseConfig_RollbackDetail) *v1pb.Plan_ChangeDatabaseConfig_RollbackDetail {
-	if d == nil {
-		return nil
-	}
-	return &v1pb.Plan_ChangeDatabaseConfig_RollbackDetail{
-		RollbackFromIssue: d.RollbackFromIssue,
-		RollbackFromTask:  d.RollbackFromIssue,
-	}
-}
-
-func convertToPlanSpecChangeDatabaseConfigType(t storepb.PlanConfig_ChangeDatabaseConfig_Type) v1pb.Plan_ChangeDatabaseConfig_Type {
-	switch t {
-	case storepb.PlanConfig_ChangeDatabaseConfig_TYPE_UNSPECIFIED:
-		return v1pb.Plan_ChangeDatabaseConfig_TYPE_UNSPECIFIED
-	case storepb.PlanConfig_ChangeDatabaseConfig_BASELINE:
-		return v1pb.Plan_ChangeDatabaseConfig_BASELINE
-	case storepb.PlanConfig_ChangeDatabaseConfig_MIGRATE:
-		return v1pb.Plan_ChangeDatabaseConfig_MIGRATE
-	case storepb.PlanConfig_ChangeDatabaseConfig_MIGRATE_SDL:
-		return v1pb.Plan_ChangeDatabaseConfig_MIGRATE_SDL
-	case storepb.PlanConfig_ChangeDatabaseConfig_MIGRATE_GHOST:
-		return v1pb.Plan_ChangeDatabaseConfig_MIGRATE_GHOST
-	case storepb.PlanConfig_ChangeDatabaseConfig_BRANCH:
-		return v1pb.Plan_ChangeDatabaseConfig_BRANCH
-	case storepb.PlanConfig_ChangeDatabaseConfig_DATA:
-		return v1pb.Plan_ChangeDatabaseConfig_DATA
-	default:
-		return v1pb.Plan_ChangeDatabaseConfig_TYPE_UNSPECIFIED
-	}
-}
-
-func convertToPlanSpecRestoreDatabaseConfig(config *storepb.PlanConfig_Spec_RestoreDatabaseConfig) *v1pb.Plan_Spec_RestoreDatabaseConfig {
-	c := config.RestoreDatabaseConfig
-	v1Config := &v1pb.Plan_Spec_RestoreDatabaseConfig{
-		RestoreDatabaseConfig: &v1pb.Plan_RestoreDatabaseConfig{
-			Target: c.Target,
-		},
-	}
-	switch source := c.Source.(type) {
-	case *storepb.PlanConfig_RestoreDatabaseConfig_Backup:
-		v1Config.RestoreDatabaseConfig.Source = &v1pb.Plan_RestoreDatabaseConfig_Backup{
-			Backup: source.Backup,
-		}
-	case *storepb.PlanConfig_RestoreDatabaseConfig_PointInTime:
-		v1Config.RestoreDatabaseConfig.Source = &v1pb.Plan_RestoreDatabaseConfig_PointInTime{
-			PointInTime: source.PointInTime,
-		}
-	}
-
-	v1Config.RestoreDatabaseConfig.CreateDatabaseConfig = convertToPlanCreateDatabaseConfig(c.CreateDatabaseConfig)
-	return v1Config
-}
-
-func convertPlanSteps(steps []*v1pb.Plan_Step) []*storepb.PlanConfig_Step {
-	storeSteps := make([]*storepb.PlanConfig_Step, len(steps))
-	for i := range steps {
-		storeSteps[i] = convertPlanStep(steps[i])
-	}
-	return storeSteps
-}
-
-func convertPlanStep(step *v1pb.Plan_Step) *storepb.PlanConfig_Step {
-	return &storepb.PlanConfig_Step{
-		Specs: convertPlanSpecs(step.Specs),
-	}
-}
-
-func convertPlanSpecs(specs []*v1pb.Plan_Spec) []*storepb.PlanConfig_Spec {
-	storeSpecs := make([]*storepb.PlanConfig_Spec, len(specs))
-	for i := range specs {
-		storeSpecs[i] = convertPlanSpec(specs[i])
-	}
-	return storeSpecs
-}
-
-func convertPlanSpec(spec *v1pb.Plan_Spec) *storepb.PlanConfig_Spec {
-	storeSpec := &storepb.PlanConfig_Spec{
-		EarliestAllowedTime: spec.EarliestAllowedTime,
-		Id:                  spec.Id,
-	}
-
-	switch v := spec.Config.(type) {
-	case *v1pb.Plan_Spec_CreateDatabaseConfig:
-		storeSpec.Config = convertPlanSpecCreateDatabaseConfig(v)
-	case *v1pb.Plan_Spec_ChangeDatabaseConfig:
-		storeSpec.Config = convertPlanSpecChangeDatabaseConfig(v)
-	case *v1pb.Plan_Spec_RestoreDatabaseConfig:
-		storeSpec.Config = convertPlanSpecRestoreDatabaseConfig(v)
-	}
-	return storeSpec
-}
-
-func convertPlanSpecCreateDatabaseConfig(config *v1pb.Plan_Spec_CreateDatabaseConfig) *storepb.PlanConfig_Spec_CreateDatabaseConfig {
-	c := config.CreateDatabaseConfig
-	return &storepb.PlanConfig_Spec_CreateDatabaseConfig{
-		CreateDatabaseConfig: convertPlanConfigCreateDatabaseConfig(c),
-	}
-}
-
-func convertPlanConfigCreateDatabaseConfig(c *v1pb.Plan_CreateDatabaseConfig) *storepb.PlanConfig_CreateDatabaseConfig {
-	return &storepb.PlanConfig_CreateDatabaseConfig{
-		Target:       c.Target,
-		Database:     c.Database,
-		Table:        c.Table,
-		CharacterSet: c.CharacterSet,
-		Collation:    c.Collation,
-		Cluster:      c.Cluster,
-		Owner:        c.Owner,
-		Backup:       c.Backup,
-		Labels:       c.Labels,
-	}
-}
-
-func convertPlanSpecChangeDatabaseConfig(config *v1pb.Plan_Spec_ChangeDatabaseConfig) *storepb.PlanConfig_Spec_ChangeDatabaseConfig {
-	c := config.ChangeDatabaseConfig
-	return &storepb.PlanConfig_Spec_ChangeDatabaseConfig{
-		ChangeDatabaseConfig: &storepb.PlanConfig_ChangeDatabaseConfig{
-			Target:          c.Target,
-			Sheet:           c.Sheet,
-			Type:            storepb.PlanConfig_ChangeDatabaseConfig_Type(c.Type),
-			SchemaVersion:   c.SchemaVersion,
-			RollbackEnabled: c.RollbackEnabled,
-		},
-	}
-}
-
-func convertPlanSpecRestoreDatabaseConfig(config *v1pb.Plan_Spec_RestoreDatabaseConfig) *storepb.PlanConfig_Spec_RestoreDatabaseConfig {
-	c := config.RestoreDatabaseConfig
-	storeConfig := &storepb.PlanConfig_Spec_RestoreDatabaseConfig{
-		RestoreDatabaseConfig: &storepb.PlanConfig_RestoreDatabaseConfig{
-			Target: c.Target,
-		},
-	}
-	switch source := c.Source.(type) {
-	case *v1pb.Plan_RestoreDatabaseConfig_Backup:
-		storeConfig.RestoreDatabaseConfig.Source = &storepb.PlanConfig_RestoreDatabaseConfig_Backup{
-			Backup: source.Backup,
-		}
-	case *v1pb.Plan_RestoreDatabaseConfig_PointInTime:
-		storeConfig.RestoreDatabaseConfig.Source = &storepb.PlanConfig_RestoreDatabaseConfig_PointInTime{
-			PointInTime: source.PointInTime,
-		}
-	}
-	// c.CreateDatabaseConfig is defined as optional in proto
-	// so we need to test if it's nil
-	if c.CreateDatabaseConfig != nil {
-		storeConfig.RestoreDatabaseConfig.CreateDatabaseConfig = convertPlanConfigCreateDatabaseConfig(c.CreateDatabaseConfig)
-	}
-	return storeConfig
-}
-
-// checkCharacterSetCollationOwner checks if the character set, collation and owner are legal according to the dbType.
-func checkCharacterSetCollationOwner(dbType db.Type, characterSet, collation, owner string) error {
-	switch dbType {
-	case db.Spanner:
-		// Spanner does not support character set and collation at the database level.
-		if characterSet != "" {
-			return errors.Errorf("Spanner does not support character set, but got %s", characterSet)
-		}
-		if collation != "" {
-			return errors.Errorf("Spanner does not support collation, but got %s", collation)
-		}
-	case db.ClickHouse:
-		// ClickHouse does not support character set and collation at the database level.
-		if characterSet != "" {
-			return errors.Errorf("ClickHouse does not support character set, but got %s", characterSet)
-		}
-		if collation != "" {
-			return errors.Errorf("ClickHouse does not support collation, but got %s", collation)
-		}
-	case db.Snowflake:
-		if characterSet != "" {
-			return errors.Errorf("Snowflake does not support character set, but got %s", characterSet)
-		}
-		if collation != "" {
-			return errors.Errorf("Snowflake does not support collation, but got %s", collation)
-		}
-	case db.Postgres:
-		if owner == "" {
-			return errors.Errorf("database owner is required for PostgreSQL")
-		}
-	case db.Redshift:
-		if owner == "" {
-			return errors.Errorf("database owner is required for Redshift")
-		}
-	case db.SQLite, db.MongoDB, db.MSSQL:
-		// no-op.
-	default:
-		if characterSet == "" {
-			return errors.Errorf("character set missing for %s", string(dbType))
-		}
-		// For postgres, we don't explicitly specify a default since the default might be UNSET (denoted by "C").
-		// If that's the case, setting an explicit default such as "en_US.UTF-8" might fail if the instance doesn't
-		// install it.
-		if collation == "" {
-			return errors.Errorf("collation missing for %s", string(dbType))
-		}
-	}
-	return nil
-}
-
-// convertDatabaseLabels converts the map[string]string labels to []*api.DatabaseLabel JSON string.
-func convertDatabaseLabels(labelsMap map[string]string) (string, error) {
-	if len(labelsMap) == 0 {
-		return "", nil
-	}
-	// For scalability, each database can have up to four labels for now.
-	if len(labelsMap) > api.DatabaseLabelSizeMax {
-		return "", errors.Errorf("database labels are up to a maximum of %d", api.DatabaseLabelSizeMax)
-	}
-	var labels []*api.DatabaseLabel
-	for k, v := range labelsMap {
-		labels = append(labels, &api.DatabaseLabel{
-			Key:   k,
-			Value: v,
-		})
-	}
-	labelsJSON, err := json.Marshal(labels)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to marshal labels json")
-	}
-	return string(labelsJSON), nil
-}
-
-func getCreateDatabaseStatement(dbType db.Type, c *storepb.PlanConfig_CreateDatabaseConfig, databaseName, adminDatasourceUser string) (string, error) {
-	var stmt string
-	switch dbType {
-	case db.MySQL, db.TiDB, db.MariaDB, db.OceanBase:
-		return fmt.Sprintf("CREATE DATABASE `%s` CHARACTER SET %s COLLATE %s;", databaseName, c.CharacterSet, c.Collation), nil
-	case db.MSSQL:
-		return fmt.Sprintf(`CREATE DATABASE "%s";`, databaseName), nil
-	case db.Postgres:
-		// On Cloud RDS, the data source role isn't the actual superuser with sudo privilege.
-		// We need to grant the database owner role to the data source admin so that Bytebase can have permission for the database using the data source admin.
-		if adminDatasourceUser != "" && c.Owner != adminDatasourceUser {
-			stmt = fmt.Sprintf("GRANT \"%s\" TO \"%s\";\n", c.Owner, adminDatasourceUser)
-		}
-		if c.Collation == "" {
-			stmt = fmt.Sprintf("%sCREATE DATABASE \"%s\" ENCODING %q;", stmt, databaseName, c.CharacterSet)
-		} else {
-			stmt = fmt.Sprintf("%sCREATE DATABASE \"%s\" ENCODING %q LC_COLLATE %q;", stmt, databaseName, c.CharacterSet, c.Collation)
-		}
-		// Set the database owner.
-		// We didn't use CREATE DATABASE WITH OWNER because RDS requires the current role to be a member of the database owner.
-		// However, people can still use ALTER DATABASE to change the owner afterwards.
-		// Error string below:
-		// query: CREATE DATABASE h1 WITH OWNER hello;
-		// ERROR:  must be member of role "hello"
-		//
-		// For tenant project, the schema for the newly created database will belong to the same owner.
-		// TODO(d): alter schema "public" owner to the database owner.
-		return fmt.Sprintf("%s\nALTER DATABASE \"%s\" OWNER TO \"%s\";", stmt, databaseName, c.Owner), nil
-	case db.ClickHouse:
-		clusterPart := ""
-		if c.Cluster != "" {
-			clusterPart = fmt.Sprintf(" ON CLUSTER `%s`", c.Cluster)
-		}
-		return fmt.Sprintf("CREATE DATABASE `%s`%s;", databaseName, clusterPart), nil
-	case db.Snowflake:
-		return fmt.Sprintf("CREATE DATABASE %s;", databaseName), nil
-	case db.SQLite:
-		// This is a fake CREATE DATABASE and USE statement since a single SQLite file represents a database. Engine driver will recognize it and establish a connection to create the sqlite file representing the database.
-		return fmt.Sprintf("CREATE DATABASE '%s';", databaseName), nil
-	case db.MongoDB:
-		// We just run createCollection in mongosh instead of execute `use <database>` first, because we execute the
-		// mongodb statement in mongosh with --file flag, and it doesn't support `use <database>` statement in the file.
-		// And we pass the database name to Bytebase engine driver, which will be used to build the connection string.
-		return fmt.Sprintf(`db.createCollection("%s");`, c.Table), nil
-	case db.Spanner:
-		return fmt.Sprintf("CREATE DATABASE %s;", databaseName), nil
-	case db.Oracle:
-		return fmt.Sprintf("CREATE DATABASE %s;", databaseName), nil
-	case db.Redshift:
-		options := make(map[string]string)
-		if adminDatasourceUser != "" && c.Owner != adminDatasourceUser {
-			options["OWNER"] = fmt.Sprintf("%q", c.Owner)
-		}
-		stmt := fmt.Sprintf("CREATE DATABASE \"%s\"", databaseName)
-		if len(options) > 0 {
-			list := make([]string, 0, len(options))
-			for k, v := range options {
-				list = append(list, fmt.Sprintf("%s=%s", k, v))
-			}
-			stmt = fmt.Sprintf("%s WITH\n\t%s", stmt, strings.Join(list, "\n\t"))
-		}
-		return fmt.Sprintf("%s;", stmt), nil
-	}
-	return "", errors.Errorf("unsupported database type %s", dbType)
-}
-
 func (s *RolloutService) createPipeline(ctx context.Context, project *store.ProjectMessage, pipelineCreate *store.PipelineMessage, creatorID int) (*store.PipelineMessage, error) {
 	pipelineCreated, err := s.store.CreatePipelineV2(ctx, &store.PipelineMessage{
 		Name:      pipelineCreate.Name,
@@ -2950,24 +928,6 @@ func (s *RolloutService) createPipeline(ctx context.Context, project *store.Proj
 	}
 
 	return pipelineCreated, nil
-}
-
-func getResourceNameForSheet(project *store.ProjectMessage, sheetUID int) string {
-	return fmt.Sprintf("%s%s/%s%d", common.ProjectNamePrefix, project.ResourceID, common.SheetIDPrefix, sheetUID)
-}
-
-func getOrDefaultSchemaVersion(v string) string {
-	if v != "" {
-		return v
-	}
-	return common.DefaultMigrationVersion()
-}
-
-func getOrDefaultSchemaVersionWithSuffix(v string, suffix string) string {
-	if v != "" {
-		return v + suffix
-	}
-	return common.DefaultMigrationVersion() + suffix
 }
 
 // canUserRunStageTasks returns if a user can run the tasks in a stage.
